Moved interfaces from the backend folder into the extension folder.
