+++ extension/algorithms/dataConversion.ts
@@ -1,4 +1,4 @@
-import { DisplayNode, State } from "../backend/interfaces";
+import { DisplayNode, State } from "../interfaces";
 const { Traverse } = require("./dataTraversal");
 
 // Export as module
