+++ extension/backend/initialHook.ts
@@ -1,4 +1,4 @@
-import { State, DisplayNode } from './interfaces';
+import { State, DisplayNode } from '../interfaces';
 const { fiberNodeToTree } = require('../algorithms/dataConversion');
 const circle = require('circular');
